requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
torch==1.1.0
torchvision==0.2.2.post3
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
